requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
arrow==1.3.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
binaryornot==0.4.4
Brotli==1.1.0
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmdstanpy==1.2.5
contourpy==1.3.3
cookiecutter==2.6.0
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
Django==5.2.5
django-auditlog==3.2.1
django-browser-reload==1.18.0
django-cid==3.0
django-filter==25.1
django-oauth-toolkit==3.0.1
django-redis==6.0.0
django-tailwind==4.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
Faker==37.4.2
fonttools==4.59.1
frozenlist==1.7.0
holidays==0.79
honcho==2.0.0
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
Jinja2==3.1.6
jwcrypto==1.5.6
kiwisolver==1.4.9
kombu==5.5.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
multidict==6.6.4
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
prophet==1.1.7
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==6.3.0
requests==2.32.4
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
stanio==0.5.1
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
twilio==9.7.0
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.9.0
yarl==1.20.1
zopfli==0.2.3.post1
